requirements/requirements.txt
django==2.2.10
djangorestframework==3.9.2
django-environ==0.4.4
psycopg2-binary==2.7.4

drf-yasg==1.17.1
